requirements.txt
aiogram==3.12.0
langchain
chromadb
langchain-gigachat
requests
python-dotenv
pydantic-settings
pytest==7.4.3
pytest-asyncio==0.21.1
langsmith == 0.2.4
langchain-mistralai==0.2.4